requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0.post1
httpx==0.27.2
paddleocr==2.7.0.3
paddlepaddle==2.6.2
pdf2image==1.17.0
pillow==10.3.0
numpy==1.26.4
